@@ -1,13 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
         <IsPackable>false</IsPackable>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
+        <PackageReference Include="System.Linq.Async" Version="5.0.0"/>
         <ProjectReference Include="..\..\src\RESTable.AspNetCore\RESTable.AspNetCore.csproj" />
         <ProjectReference Include="..\..\src\RESTable.Excel\RESTable.Excel.csproj" />
         <ProjectReference Include="..\..\src\RESTable.OData\RESTable.OData.csproj" />
