@@ -47,17 +47,19 @@ namespace RESTable
         /// <summary>
         /// Deserializes the body to an IEnumerable of entities of the given type
         /// </summary>
-        public IAsyncEnumerable<T> Deserialize<T>()
+        public IAsyncEnumerable<T> Deserialize<T>(CancellationToken cancellationToken = new())
         {
             if (IsClosed) throw new ObjectDisposedException(nameof(Stream));
             try
             {
-                return ContentTypeProvider.DeserializeCollection<T>(Stream);
+                return ContentTypeProvider.DeserializeCollection<T>(Stream, cancellationToken);
             }
             finally
             {
                 if (CanSeek)
+                {
                     Stream.Rewind();
+                }
             }
         }
 
