@@ -12,13 +12,9 @@ namespace RESTable.ContentTypeProviders
         void Populate(object target, string json);
         void Populate(object target, JsonElement json);
 
-        IJsonWriter GetJsonWriter(TextWriter textWriter);
-
-        Task SerializeAsync(IJsonWriter jsonWriter, object value);
+        string Serialize(object value, bool? prettyPrint = null, bool ignoreNulls = false);
         Task SerializeAsync<T>(Stream stream, T entity, bool? prettyPrint = null, bool ignoreNulls = false, CancellationToken cancellationToken = new());
-        
-        string Serialize(object value, bool? prettyPrint = null, bool ignoreNulls = false); 
-        ValueTask<long> SerializeCollectionAsync<T>(IJsonWriter textWriter, IAsyncEnumerable<T> collectionObject, CancellationToken cancellationToken) where T : class;
+        ValueTask<long> SerializeCollectionAsync<T>(Utf8JsonWriter writer, IAsyncEnumerable<T> collectionObject, CancellationToken cancellationToken) where T : class;
 
         T? Deserialize<T>(byte[] bytes);
         T? Deserialize<T>(byte[] bytes, int offset, int count);
@@ -28,18 +24,5 @@ namespace RESTable.ContentTypeProviders
 
         ValueTask<T?> DeserializeAsync<T>(Stream stream);
         ValueTask<object?> DeserializeAsync(Stream stream, Type targetType);
-    }
-
-    public interface IJsonWriter : IDisposable
-    {
-        void StartCountObjectsWritten();
-        long StopCountObjectsWritten();
-        Task WriteStartObjectAsync(CancellationToken cancellationToken);
-        Task WritePropertyNameAsync(string status, CancellationToken cancellationToken);
-        Task WriteEndObjectAsync(CancellationToken cancellationToken);
-        Task WriteValueAsync(long value, CancellationToken cancellationToken);
-        Task WriteValueAsync(double value, CancellationToken cancellationToken);
-        Task WriteValueAsync(string value, CancellationToken cancellationToken);
-        Task WriteValueAsync(bool value, CancellationToken cancellationToken);
     }
 }
